requirements.txt
fastapi[all]
uvicorn
pyrebase4
